Refactor to move big "factory array" to its own file.
